+++ src/i18n.ts
@@ -14,7 +14,7 @@ export const defaultLang = { code: "en", label: "English" };
 const allLanguages: Language[] = [
   { code: "ar-SA", label: "العربية", rtl: true },
   { code: "bg-BG", label: "Български" },
-  { code: "ca-ES", label: "Catalan" },
+  { code: "ca-ES", label: "Català" },
   { code: "de-DE", label: "Deutsch" },
   { code: "el-GR", label: "Ελληνικά" },
   { code: "es-ES", label: "Español" },
@@ -32,6 +32,7 @@ const allLanguages: Language[] = [
   { code: "nb-NO", label: "Norsk bokmål" },
   { code: "nl-NL", label: "Nederlands" },
   { code: "nn-NO", label: "Norsk nynorsk" },
+  { code: "pa-IN", label: "ਪੰਜਾਬੀ" },
   { code: "pl-PL", label: "Polski" },
   { code: "pt-BR", label: "Português Brasileiro" },
   { code: "pt-PT", label: "Português" },